fixed docs, renaming
