+++ yarn-project/pxe/src/entrypoints/client/bundle/index.ts
@@ -2,4 +2,4 @@ export * from '../../../pxe_service/index.js';
 export * from '../../../config/index.js';
 export * from '../../../storage/index.js';
 export * from './utils.js';
-export { PXEOracleInterface } from '../../../pxe_oracle_interface/index.js';
+export { PXEOracleInterface } from '../../../pxe_oracle_interface/pxe_oracle_interface.js';
