+++ yarn-project/pxe/src/entrypoints/server/index.ts
@@ -3,4 +3,4 @@ export * from '../../pxe_http/index.js';
 export * from '../../config/index.js';
 export * from '../../storage/index.js';
 export * from './utils.js';
-export { PXEOracleInterface } from '../../pxe_oracle_interface/index.js';
+export { PXEOracleInterface } from '../../pxe_oracle_interface/pxe_oracle_interface.js';
