+++ yarn-project/pxe/src/pxe_oracle_interface/pxe_oracle_interface.test.ts
@@ -26,7 +26,7 @@ import { ContractDataProvider } from '../storage/contract_data_provider/contract
 import { NoteDataProvider } from '../storage/note_data_provider/note_data_provider.js';
 import { SyncDataProvider } from '../storage/sync_data_provider/sync_data_provider.js';
 import { TaggingDataProvider } from '../storage/tagging_data_provider/tagging_data_provider.js';
-import { PXEOracleInterface } from './index.js';
+import { PXEOracleInterface } from './pxe_oracle_interface.js';
 import { WINDOW_HALF_SIZE } from './tagging_utils.js';
 
 const TXS_PER_BLOCK = 4;
